refactor(tui): project.py を project_modal.py に統合する
tui/project.py は redi.api.project と紛らわしく、隣の tui/project_modal.py と
どちらに何があるか名前から分からなかった。中身は modal を開く/切り替える操作
だけなので、描画を持つ project_modal.py に寄せてプロジェクト切替を 1 ファイルに
まとめる。
